% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_data.R
\name{new_data}
\alias{new_data}
\title{Resample a data frame}
\usage{
new_data(Data, Column_Name = NA)
}
\arguments{
\item{Data}{A data frame}

\item{Column_Name}{The name of the column within the data frame that is to be
resampled}
}
\value{
The same data frame provided, but with the specified column
randomized.
}
\description{
This function resamples a column in a data frame. It first removes all NA's
from the column, the samples that column without replacement. The resampled
column then replaces the original column in the data frame.
}
\keyword{internal}
