% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_test.R
\name{perm_test}
\alias{perm_test}
\title{Perform a permutation test}
\usage{
perm_test(Model_Object, Data, Variable, Replication)
}
\arguments{
\item{Model_Object}{A statistical model object (change name).}

\item{Data}{The data that the model is built from.}

\item{Variable}{The variable requiring the permutation test.}

\item{Replication}{The number of simulations in the permutation test.}
}
\value{
A list of two items. The first is a data frame of results of the
permutation test. The second is a histogram displaying the sampling
distribution of the simulated test statistics, with a red line displaying the
test statistic of the original (non-randomized) data.
}
\description{
This function performs a permutation test on a model object. Given
a model object, a data frame, and a variable. It makes use of
the new_data function to resample a variable in a data frame. Then, it
generates a specified amount of replicated data frames and fits the model to
each one in order to perform a permutation test.
}
