% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_test_LR.R
\name{perm_test_LR}
\alias{perm_test_LR}
\title{Perform a permutation test}
\usage{
perm_test_LR(
  Model_Object,
  Data,
  Variable,
  Replication,
  UseAllAvailableCores = TRUE
)
}
\arguments{
\item{Model_Object}{A statistical model object (change name).}

\item{Data}{The data that the model is built from.}

\item{Variable}{The variable requiring the permutation test.}

\item{Replication}{The number of simulations in the permutation test.}
}
\value{
A list of two items. The first is a data frame of results of the
permutation test. The second is a histogram displaying the sampling
distribution of the simulated test statistics, with a red line displaying the
test statistic of the original (non-randomized) data.
}
\description{
This function performs a permutation test on a model object. It makes use of
the new_data function to resample a variable in a data frame. Then, it
generates a specified amount of replicated data frames and fits the supplied
model to each one. The permulation test is based on the f values generated
from a likelyhood ratio test. The null model is determined to be the
Model_Object, updated to exclude the supplied variable.
}
