% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermTest_Contrast.R
\name{permTest_Contrast}
\alias{permTest_Contrast}
\title{Perform a permutation test on an interaction using the contrast specific in
the model output.}
\usage{
permTest_Contrast(
  Model_Object,
  Test_Parameter,
  Randomize_Variables,
  Test_Statistic,
  Replication,
  UseAllAvailableCores = TRUE,
  OutputData = FALSE
)
}
\arguments{
\item{Model_Object}{A statistical model object.}

\item{Test_Parameter}{A character string indicating the parameter that the
permutation test is to be performed on. For example, 'var1:var2'. If testing
the effect of a categorical variable or interaction, see the output of summary
of the model to determine what parameter the permutation test will be built on.}

\item{Randomize_Variables}{A character vector indicating the variables that
are to be randomized in the permutation test}

\item{Test_Statistic}{A Character string indicating the desired test statistic
to conduct the permutation test.This can be an f value, chi-squared value,
p-value...etc, depending on what test statistics are available for your model
object. These are different for lm, glm, glmmTMB, etc.
using the 'anova' function and choose one of the output parameters.}

\item{Replication}{The number of simulations in the permutation test.}

\item{OutputData}{Should the simulated test statistics be outputted ?}

\item{Data}{The data that the model is built from.}
}
\value{
A list of two items. The first is a data frame of results of the
permutation test. The second is a histogram displaying the sampling
distribution of the simulated test statistics, with a red line displaying the
test statistic of the original (non-randomized) data. If OutputData=T, then
a vector of simulated test statistics will also be included in the output list.
}
\description{
This function performs a permutation test on a model object. The permutation
test is based on the test statistic from the summary of the model, which are
typically contrasts. One must specify the variables to be randomized, the test
parameter (i.e. the variable of interest), and the amount of replication to be
performed. This function works on lm, glm, lmer, glmer and glmmTMB classes.
}
