% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permTest_LR.R
\name{permTest_LR}
\alias{permTest_LR}
\title{Perform a permutation test using the likelyhood ratio test}
\usage{
permTest_LR(
  Model_Object,
  Variable,
  Test_Statistic,
  Replication,
  UseAllAvailableCores = TRUE,
  OutputData = FALSE
)
}
\arguments{
\item{Model_Object}{A statistical model object.}

\item{Variable}{A character string of the variable requiring the permutation
test is run on.}

\item{Test_Statistic}{A character string of the desired test statistic to
conduct the permutation test.}

\item{Replication}{The number of simulations in the permutation test.}

\item{OutputData}{Should the simulated test statistics be outputted ?}

\item{Data}{The data that the model is built from.}
}
\value{
A list of two items. The first is a data frame of results of the
permutation test. The second is a histogram displaying the sampling
distribution of the simulated test statistics, with a red line displaying the
test statistic of the original (non-randomized) data. If OutputData=T, then
a vector of simulated test statistics will also be included in the output list.
}
\description{
This function performs a permutation test on a model object using a likelyhood
ratio test. It works by generating a specified amount of replicated
data frames and fits the supplied model object to each one. The permulation
test is based on the test statistic output of  a likelyhood ratio test.
The null model is the Model_Object without the variable of interest. This
function will not work on interactions, to run a permutation test on an
interaction with a likelyhood ratio test, use permTest_LR_int().
}
