% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permTest_LR_int.R
\name{permTest_LR_int}
\alias{permTest_LR_int}
\title{Perform a permutation test on an interaction using the likelyhood ratio test}
\usage{
permTest_LR_int(
  Model_Object,
  Test_Parameter,
  Randomize_Variables,
  Test_Statistic,
  Replication,
  UseAllAvailableCores = TRUE,
  OutputData = FALSE
)
}
\arguments{
\item{Model_Object}{A statistical model object.}

\item{Test_Parameter}{A character string indicating the parameter that the
permutation test is to be performed on.For example, 'var1:var2', or "var1"}

\item{Randomize_Variables}{A character vector of variables that are to be
randomized in the permutation test. ex. c("var1","var2") or "var1".}

\item{Test_Statistic}{A character string indicating the desired test
statistic to conduct the permutation test. This can be an f value,
chi-squared value, p-value...etc, depending on what test statistics are
available for your model object. These are different for lm, glm, glmmTMB, etc.}

\item{Replication}{The number of simulations in the permutation test.}

\item{OutputData}{Should the simulated test statistics be outputted ?}

\item{Data}{The data that the model is built from.}
}
\value{
A list of two items. The first is a data frame of results of the
permutation test. The second is a histogram displaying the sampling
distribution of the simulated test statistics, with a red line displaying the
test statistic of the original (non-randomized) data. If OutputData=T, then
a vector of simulated test statistics will also be included in the output list.
}
\description{
This function performs a permutation test on a model object. It makes use of
the new_data function to resample a variable in a data frame. Then, it
generates a specified amount of replicated data frames and fits the supplied
model to each one. The permulation test is based on the test statistic generated
from a likelyhood ratio test. The null model is determined to be the
Model_Object, updated to exclude the supplied variable.
}
