% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_extract.R
\name{model_extract}
\alias{model_extract}
\title{Extract Focal Parameter from Model}
\usage{
model_extract(Data, Model_Object, Null_Model, ...)
}
\arguments{
\item{Data}{The data that the model is built from.}

\item{Model_Object}{A statistical model object (change name)}
}
\value{
The desired test statistic
}
\description{
This function extracts the desired test statistic from a model object. This
test statistic will provide the foundation of the permutation test. During the
permutation test, only this parameter will be stored.
}
